+++ packages/rocketh-deploy/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rocketh/deploy
 
+## 0.11.14
+
+### Patch Changes
+
+- LinkedData + remove auto-json-convertion
+- Updated dependencies
+  - rocketh@0.11.14
+
 ## 0.11.13
 
 ### Patch Changes