+++ packages/rocketh-diamond/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @rocketh/diamond
 
+## 0.11.16
+
+### Patch Changes
+
+- LinkedData + remove auto-json-convertion
+- Updated dependencies
+  - @rocketh/deploy@0.11.14
+  - rocketh@0.11.14
+  - @rocketh/read-execute@0.11.14
+
 ## 0.11.15
 
 ### Patch Changes