+++ packages/rocketh-read-execute/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rocketh/deploy
 
+## 0.11.14
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.11.14
+
 ## 0.11.13
 
 ### Patch Changes