+++ packages/rocketh-router/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rocketh/router
 
+## 0.11.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @rocketh/deploy@0.11.14
+  - rocketh@0.11.14
+
 ## 0.11.13
 
 ### Patch Changes